Split game extractor class
